⚡Core: Moved Server Pages JS layout
