+++ src/Directory/Directory/Pages/js/layouts/components/Logo.js
@@ -1,0 +1,13 @@
+import React from "react";
+import { Image, Link } from "@chakra-ui/core";
+
+const Logo = () => (
+  <Link href="/">
+    <Image
+      src="/images/UKCRC_TDCC_logo.jpg"
+      alt="UKCRC Tissue Directory and Coordination Centre Logo"
+    />
+  </Link>
+);
+
+export default Logo;
