+++ src/Directory/Directory/Pages/js/pages/Account/ForgotPassword/ForgotPasswordResult.js
@@ -1,5 +1,5 @@
 import React from "react";
-import Layout from "@/components/Layout";
+import Layout from "@/layouts/Clean";
 import BasicAlert from "@/components/BasicAlert";
 import { AlertDescription, Text, Box, Link } from "@chakra-ui/core";
