+++ src/Directory/Directory/Pages/js/pages/Account/Login/Login.js
@@ -1,6 +1,6 @@
 import React from "react";
 import { Flex, Link } from "@chakra-ui/core";
-import Layout from "@/components/Layout";
+import Layout from "@/layouts/Clean";
 import UnconfirmedAccountFound from "@/components/UnconfirmedAccountFound";
 import ModelValidationSummary from "@/components/ModelValidationSummary";
 import TryThisAlert from "@/components/TryThisAlert";
