+++ src/Directory/Directory/Pages/js/pages/Account/Register/RegisterResult.js
@@ -1,5 +1,5 @@
 import React from "react";
-import Layout from "@/components/Layout";
+import Layout from "@/layouts/Clean";
 import ConfirmationSent from "@/components/ConfirmationSent";
 
 const RegisterResult = ({ Email }) => {
